Attempt at fixing challenges in 01-intro-to-r.Rmd.
I looked at an episode where challenges were rendering properly and at this one, and made changes based on the differences I saw. **These need to be tested before they are accepted.**

I also removed some code blocks that just duplicated the challenges, because again, their equivalent were not present in the episode that was working.